test: define common sets of tests for each FHIR version (WIP)
