+++ r4/src/test/java/io/github/jy95/fds/r4/translators/BoundsPeriodTest.java
@@ -1,21 +1,17 @@
 package io.github.jy95.fds.r4.translators;
 
+import io.github.jy95.fds.common.types.DisplayOrder;
 import io.github.jy95.fds.common.types.DosageAPI;
 import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
-import io.github.jy95.fds.common.types.DisplayOrder;
 import io.github.jy95.fds.r4.config.FDSConfigR4;
 import io.github.jy95.fds.translators.AbstractBoundsPeriodTest;
-import org.hl7.fhir.r4.model.*;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.hl7.fhir.r4.model.DateTimeType;
+import org.hl7.fhir.r4.model.Dosage;
+import org.hl7.fhir.r4.model.Period;
+import org.hl7.fhir.r4.model.Timing;
 
 import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class BoundsPeriodTest extends AbstractBoundsPeriodTest<FDSConfigR4, Dosage> {
