+++ r4/src/test/java/io/github/jy95/fds/r4/translators/BoundsRangeTest.java
@@ -1,20 +1,17 @@
 package io.github.jy95.fds.r4.translators;
 
+import io.github.jy95.fds.common.types.DisplayOrder;
 import io.github.jy95.fds.common.types.DosageAPI;
 import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
-import io.github.jy95.fds.common.types.DisplayOrder;
 import io.github.jy95.fds.r4.config.FDSConfigR4;
 import io.github.jy95.fds.translators.AbstractBoundsRangeTest;
-import org.hl7.fhir.r4.model.*;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.hl7.fhir.r4.model.Dosage;
+import org.hl7.fhir.r4.model.Quantity;
+import org.hl7.fhir.r4.model.Range;
+import org.hl7.fhir.r4.model.Timing;
 
 import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class BoundsRangeTest extends AbstractBoundsRangeTest<FDSConfigR4, Dosage> {
 
@@ -97,7 +94,7 @@ public class BoundsRangeTest extends AbstractBoundsRangeTest<FDSConfigR4, Dosage
         dosage.setTiming(timing);
         return dosage;
     }
-    
+
     @Override
     public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
         return new DosageAPIR4(FDSConfigR4.builder()
